Prevented casting of power values to ints

+++ arduino_code/jankboat/jankboat.ino
@@ -24,36 +24,36 @@ float mapFloat(float val, float inmin, float inmax, float outmin, float outmax){
 
 Servo motor1;
 void motor1_cb( const std_msgs::Float32& cmd_msg) {
-  int inputPower = cmd_msg.data; // -1 to 1
+  float inputPower = cmd_msg.data; // -1 to 1
   inputPower = constrain( inputPower, -1, 1);
-  int outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
+  float outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
   motor1.writeMicroseconds(outputPower);
   lastMessageMillis = millis();
 }
 
 Servo motor2;
 void motor2_cb( const std_msgs::Float32& cmd_msg) {
-  int inputPower = cmd_msg.data; // -1 to 1
+  float inputPower = cmd_msg.data; // -1 to 1
   inputPower = constrain( inputPower, -1, 1);
-  int outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
+  float outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
   motor2.writeMicroseconds(outputPower);
   lastMessageMillis = millis();
 }
 
 Servo motor3;
 void motor3_cb( const std_msgs::Float32& cmd_msg) {
-  int inputPower = cmd_msg.data; // -1 to 1
+  float inputPower = cmd_msg.data; // -1 to 1
   inputPower = constrain( inputPower, -1, 1);
-  int outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
+  float outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
   motor3.writeMicroseconds(outputPower);
   lastMessageMillis = millis();
 }
 
 Servo motor4;
 void motor4_cb( const std_msgs::Float32& cmd_msg) {
-  int inputPower = cmd_msg.data; // -1 to 1
+  float inputPower = cmd_msg.data; // -1 to 1
   inputPower = constrain( inputPower, -1, 1);
-  int outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
+  float outputPower = mapFloat(inputPower, -1, 1, 1000, 2000); // Convert to usuable microseconds
   motor4.writeMicroseconds(outputPower);
   lastMessageMillis = millis();
 }
